XmlAdaptedRecord: Relocate to elements package
