+++ src/main/java/seedu/address/storage/elements/XmlAdaptedRecord.java
@@ -1,4 +1,4 @@
-package seedu.address.storage;
+package seedu.address.storage.elements;
 
 import java.util.Objects;
